% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWIOT.R
\name{getWIOT}
\alias{getWIOT}
\title{Download a WIOT from the 2016 Release of the WIOD project}
\usage{
getWIOT(period = 2010, format = "wide", as.DT = TRUE,
  version = "October16")
}
\arguments{
\item{period}{the year of the WIOT}

\item{format}{a character string to specify the needed format,
either "wide" (default), "long" or "list". See Details for more
information.}

\item{as.DT}{a logical value: Should the "wide" or "long" formats
be returned as data.tables? Default is TRUE.}

\item{version}{for (possible further) revisions. Currently not
used.}
}
\value{
the function returns either a data.table (or data.frame
    when \code{as.DT = FALSE}) for the chosen formats "wide" or "long", or
    a list for format "list"
}
\description{
This function allows you to download a World Input-Output Table
(WIOT) from the 2016 release of the \href{www.wiod.org}{WIOD
project} for a given year. Additionally you can specify among three
possible formats (wide, long, list).
}
\details{
If \code{format = "wide"} is chosen, the data is returned in a
    wide data.table. This format is the format you know from the
    Excel sheets (available from the WIOD homepage).

    The format option \code{"long"} returns a reshaped/melted
    data.table.

    Finally, \code{format = "list"} returns a list containing these
    matrices/vectors:
\enumerate{

\item matrix \strong{Z}, the square (2464 x 2464) matrix of intermediates,
\item matrix \strong{F}, the final demand block (2464 x 220),
\item vector \strong{x}, gross output per country-sector,
\item vector \strong{v}, value added per country-sector,
\item matrix \strong{va.block}, the valuation block (i.e. taxes, adjustments
    for exports, international trade margins, etc.) and
\item 1-element vector \strong{year}, specifying the year of the WIOT.
}
}
\examples{
wiot.2005 <- getWIOT(period = 2005, format = "wide")

##
## get all WIOTs at once
##
## wiod.list <- list()
## for(this.year in 2000:2014) {
##   wiod.list[[as.character(this.year)]] <- getWIOT(period = this.year,
##                                                     format = "list")
##}

##
## decompose the 2005 WIOT using the WWZ decomposition
##
## library(decompr)
## wiot.2005 <- getWIOT(period = 2005, format = "list")
## industries <- 1:56
## countries <- unique(substring(names(wiot.2005$x), 1, 3))
## decomp <- decomp(x = wiot.2005$Z,
##                  y = wiot.2005$F,
##                  k = countries,
##                  i = industries,
##                  o = wiot.2005$x,
##                  V = wiot.2005$v,
##                  method = "wwz",
##                  verbose = TRUE)
}
\author{
Oliver Reiter
}
\references{
Timmer, M. P., Dietzenbacher, E., Los, B., Stehrer, R. and de
Vries, G. J. (2015),
"An Illustrated User Guide to the World Input-Output Database: the Case of Global Automotive Production". Available
\href{http://dx.doi.org/10.1111/roie.12178}{here}.
}
\seealso{
\url{http://www.wiod.org/database/wiots16}
}
\keyword{data,}
\keyword{input-output,}
\keyword{wiod}

