% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndustryClassification.R
\docType{data}
\name{WIOD_Corr}
\alias{WIOD_Corr}
\title{Correspondences to A*64 NACE Rev2 / ISIC Rev4 and Eurostat}
\format{
A list of two: a 64 x 2 and a 13 x 2 data.table.
}
\usage{
data(WIOD_Corr)
}
\description{
The WIOD 2016 Release uses an (slightly aggregated) industry
classification which is based on the NACE Rev2 / ISIC Rev4 (on
2-digit level, these two classifications coincide). The data.table
\code{WIOD_NACErev2} gives a correspondence to these commonly used
64-industries classifications.  Additionally, the WIOD uses a
different set of codes for the final demand and valuation
columns. \code{WIOD_ESTAT} gives a correspondence of those codes to
the codes used by Eurostat.
}
\examples{
data(WIOD_Corr)
}
\keyword{classification,}
\keyword{correspondence}
\keyword{data,}
\keyword{industry}
