% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndustryClassification.R
\docType{data}
\name{WIOD_IndClass}
\alias{WIOD_IndClass}
\title{Industry classification}
\format{A 69 x 3 data.table}
\usage{
data(WIOD_IndClass)
}
\description{
A data.table named \code{WIOD_IndClass} with the industry classification
used in the WIOD 2016 Release.
}
\examples{
data(WIOD_IndClass)
}
\keyword{classification}
\keyword{data,}
\keyword{industry}

