% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiod-data.R
\docType{data}
\name{wiot_reg_1995}
\alias{wiot_reg_1995}
\title{Regional WIOD data (2013 Release) for the year 1995}
\description{
A regional WIOT for year 1995 will be loaded into the
    environment. The input-output table data is contained in a list
    with the following items:

\itemize{

\item matrix \strong{inter}, the square (1435 x 1435) matrix of
intermediates,

\item matrix \strong{final}, the final demand block (1435 x 205),

\item vector \strong{output}, gross output per country-sector,

\item vector \strong{countries}, a character vector of country
names (ISO3),

\item vector \strong{industries}, a character vector of industry names
}
}
