% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wiod.R
\docType{package}
\name{wiod}
\alias{wiod}
\alias{wiod-package}
\title{World Input Output Database 1995-2011}
\description{
World Input Output Database 1995-2011
}
\references{
{Timmer, Marcel P. (ed) (2012), "The World Input-Output Database (WIOD): Contents Sources and Methods", WIOD Working Paper Number 10, downloadable at http://www.wiod.org/publications/papers/wiod10.pdf }
}
\seealso{
http://wwww.wiod.org/ http://qua.st/wiod
}

